backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.0
pymongo==4.7.2
openai==1.30.1
pydantic==2.7.1
python-dotenv==1.0.1
